Update Movie_RcSys_Project.ipynb
This is my toy project that I worked for my graduate school final project. 
In this code, I mainly deployed CNN DL network to train the dataset. 
The other explanation for this project is explained in detailed way line by line inside the Jupiter notebook.
The original book is written by korean&Chinese, for the international buddies, this notebook will be translated in english by time goes on.